chore(monomodule): refactors openai package.
